pyproject.toml
[project]
name = "claude-playground"
version = "0.1.0"
description = "Semantic word similarity finder using FastText embeddings"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=4.2",
    "fasttext-wheel>=0.9.2",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "jupytext>=1.15.0",
    "pandas>=2.0.0",
    "tqdm>=4.65.0",
    "djangorestframework>=3.14.0",
    "drf-spectacular>=0.27.0",
]
